// apps/web/src/server/db/schema.gen.ts
export type DbDateTime = string;

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface ApiKeyTable {
  apikey: string;
  can_create_tickets: Generated<number>;
  can_exec_cron: Generated<number>;
  created: DbDateTime;
  id: Generated<number>;
  ipaddr: string;
  isactive: Generated<number>;
  notes: Generated<string | null>;
  updated: DbDateTime;
}

export interface AttachmentTable {
  file_id: number;
  id: Generated<number>;
  inline: Generated<number>;
  lang: Generated<string | null>;
  name: Generated<string | null>;
  object_id: number;
  type: string;
}

export interface CannedResponseTable {
  canned_id: Generated<number>;
  created: DbDateTime;
  dept_id: Generated<number>;
  isenabled: Generated<number>;
  lang: Generated<string>;
  notes: Generated<string | null>;
  response: string;
  title: Generated<string>;
  updated: DbDateTime;
}

export interface ConfigTable {
  id: Generated<number>;
  key: string;
  namespace: string;
  updated: Generated<DbDateTime>;
  value: string;
}

export interface ContentTable {
  body: string;
  created: DbDateTime;
  id: Generated<number>;
  isactive: Generated<number>;
  name: string;
  notes: Generated<string | null>;
  type: Generated<string>;
  updated: DbDateTime;
}

export interface DepartmentTable {
  autoresp_email_id: Generated<number>;
  created: DbDateTime;
  email_id: Generated<number>;
  flags: Generated<number>;
  group_membership: Generated<number>;
  id: Generated<number>;
  ispublic: Generated<number>;
  manager_id: Generated<number>;
  message_auto_response: Generated<number>;
  name: Generated<string>;
  path: Generated<string>;
  pid: Generated<number | null>;
  schedule_id: Generated<number>;
  signature: string;
  sla_id: Generated<number>;
  ticket_auto_response: Generated<number>;
  tpl_id: Generated<number>;
  updated: DbDateTime;
}

export interface DraftTable {
  body: string;
  created: Generated<DbDateTime>;
  extra: Generated<string | null>;
  id: Generated<number>;
  namespace: Generated<string>;
  staff_id: number;
  updated: Generated<DbDateTime | null>;
}

export interface EmailTable {
  created: DbDateTime;
  dept_id: Generated<number>;
  email: Generated<string>;
  email_id: Generated<number>;
  name: Generated<string>;
  noautoresp: Generated<number>;
  notes: Generated<string | null>;
  priority_id: Generated<number>;
  topic_id: Generated<number>;
  updated: DbDateTime;
}

export interface EmailAccountTable {
  active: Generated<number>;
  allow_spoofing: Generated<number | null>;
  archivefolder: Generated<string | null>;
  auth_bk: string;
  auth_id: Generated<string | null>;
  created: DbDateTime;
  email_id: number;
  encryption: Generated<"AUTO" | "NONE" | "SSL">;
  fetchfreq: Generated<number>;
  fetchmax: Generated<number | null>;
  folder: Generated<string | null>;
  host: Generated<string>;
  id: Generated<number>;
  last_activity: Generated<DbDateTime | null>;
  last_error: Generated<DbDateTime | null>;
  last_error_msg: Generated<string | null>;
  num_errors: Generated<number>;
  port: number;
  postfetch: Generated<"archive" | "delete" | "nothing">;
  protocol: Generated<"IMAP" | "OTHER" | "POP" | "SMTP">;
  type: Generated<"mailbox" | "smtp">;
  updated: Generated<DbDateTime>;
}

export interface EmailTemplateTable {
  body: string;
  code_name: string;
  created: DbDateTime;
  id: Generated<number>;
  notes: Generated<string | null>;
  subject: Generated<string>;
  tpl_id: number;
  updated: DbDateTime;
}

export interface EmailTemplateGroupTable {
  created: DbDateTime;
  isactive: Generated<number>;
  lang: Generated<string>;
  name: Generated<string>;
  notes: Generated<string | null>;
  tpl_id: Generated<number>;
  updated: DbDateTime;
}

export interface EventTable {
  description: Generated<string | null>;
  id: Generated<number>;
  name: string;
}

export interface FaqTable {
  answer: string;
  category_id: Generated<number>;
  created: DbDateTime;
  faq_id: Generated<number>;
  ispublished: Generated<number>;
  keywords: Generated<string | null>;
  notes: Generated<string | null>;
  question: string;
  updated: DbDateTime;
}

export interface FaqCategoryTable {
  category_id: Generated<number>;
  category_pid: Generated<number | null>;
  created: DbDateTime;
  description: string;
  ispublic: Generated<number>;
  name: Generated<string | null>;
  notes: string;
  updated: DbDateTime;
}

export interface FaqTopicTable {
  faq_id: number;
  topic_id: number;
}

export interface FileTable {
  attrs: Generated<string | null>;
  bk: Generated<string>;
  created: DbDateTime;
  ft: Generated<string>;
  id: Generated<number>;
  key: string;
  name: Generated<string>;
  signature: string;
  size: Generated<number>;
  type: Generated<string>;
}

export interface FileChunkTable {
  chunk_id: number;
  file_id: number;
  filedata: Buffer;
}

export interface FilterTable {
  created: DbDateTime;
  email_id: Generated<number>;
  execorder: Generated<number>;
  flags: Generated<number | null>;
  id: Generated<number>;
  isactive: Generated<number>;
  match_all_rules: Generated<number>;
  name: Generated<string>;
  notes: Generated<string | null>;
  status: Generated<number>;
  stop_onmatch: Generated<number>;
  target: Generated<"Any" | "API" | "Email" | "Web">;
  updated: DbDateTime;
}

export interface FilterActionTable {
  configuration: Generated<string | null>;
  filter_id: number;
  id: Generated<number>;
  sort: Generated<number>;
  type: string;
  updated: DbDateTime;
}

export interface FilterRuleTable {
  created: DbDateTime;
  filter_id: Generated<number>;
  how: "contains" | "dn_contain" | "ends" | "equal" | "match" | "not_equal" | "not_match" | "starts";
  id: Generated<number>;
  isactive: Generated<number>;
  notes: string;
  updated: DbDateTime;
  val: string;
  what: string;
}

export interface FormTable {
  created: DbDateTime;
  flags: Generated<number>;
  id: Generated<number>;
  instructions: Generated<string | null>;
  name: Generated<string>;
  notes: Generated<string | null>;
  pid: Generated<number | null>;
  title: string;
  type: Generated<string>;
  updated: DbDateTime;
}

export interface FormEntryTable {
  created: DbDateTime;
  extra: Generated<string | null>;
  form_id: number;
  id: Generated<number>;
  object_id: Generated<number | null>;
  object_type: Generated<string>;
  sort: Generated<number>;
  updated: DbDateTime;
}

export interface FormEntryValuesTable {
  entry_id: number;
  field_id: number;
  value: Generated<string | null>;
  value_id: Generated<number | null>;
}

export interface FormFieldTable {
  configuration: Generated<string | null>;
  created: DbDateTime;
  flags: Generated<number | null>;
  form_id: number;
  hint: Generated<string | null>;
  id: Generated<number>;
  label: string;
  name: string;
  sort: number;
  type: Generated<string>;
  updated: DbDateTime;
}

export interface GroupTable {
  created: DbDateTime;
  flags: Generated<number>;
  id: Generated<number>;
  name: Generated<string>;
  notes: Generated<string | null>;
  role_id: number;
  updated: DbDateTime;
}

export interface HelpTopicTable {
  created: DbDateTime;
  dept_id: Generated<number>;
  flags: Generated<number | null>;
  ispublic: Generated<number>;
  noautoresp: Generated<number>;
  notes: Generated<string | null>;
  number_format: Generated<string | null>;
  page_id: Generated<number>;
  priority_id: Generated<number>;
  sequence_id: Generated<number>;
  sla_id: Generated<number>;
  sort: Generated<number>;
  staff_id: Generated<number>;
  status_id: Generated<number>;
  team_id: Generated<number>;
  topic: Generated<string>;
  topic_id: Generated<number>;
  topic_pid: Generated<number>;
  updated: DbDateTime;
}

export interface HelpTopicFormTable {
  extra: Generated<string | null>;
  form_id: Generated<number>;
  id: Generated<number>;
  sort: Generated<number>;
  topic_id: Generated<number>;
}

export interface ListTable {
  configuration: Generated<string>;
  created: DbDateTime;
  id: Generated<number>;
  masks: Generated<number>;
  name: string;
  name_plural: Generated<string | null>;
  notes: Generated<string | null>;
  sort_mode: Generated<"-Alpha" | "Alpha" | "SortCol">;
  type: Generated<string | null>;
  updated: DbDateTime;
}

export interface ListItemsTable {
  extra: Generated<string | null>;
  id: Generated<number>;
  list_id: Generated<number | null>;
  properties: Generated<string | null>;
  sort: Generated<number>;
  status: Generated<number>;
  value: string;
}

export interface LockTable {
  code: Generated<string | null>;
  created: DbDateTime;
  expire: Generated<DbDateTime | null>;
  lock_id: Generated<number>;
  staff_id: Generated<number>;
}

export interface NoteTable {
  body: Generated<string | null>;
  created: Generated<DbDateTime>;
  ext_id: Generated<string | null>;
  id: Generated<number>;
  pid: Generated<number | null>;
  sort: Generated<number>;
  staff_id: Generated<number>;
  status: Generated<number>;
  updated: Generated<DbDateTime>;
}

export interface OrganizationTable {
  created: Generated<DbDateTime | null>;
  domain: Generated<string>;
  extra: Generated<string | null>;
  id: Generated<number>;
  manager: Generated<string>;
  name: Generated<string>;
  status: Generated<number>;
  updated: Generated<DbDateTime | null>;
}

export interface OrganizationCdataTable {
  address: Generated<string | null>;
  name: Generated<string | null>;
  notes: Generated<string | null>;
  org_id: number;
  phone: Generated<string | null>;
  website: Generated<string | null>;
}

export interface PluginTable {
  id: Generated<number>;
  install_path: string;
  installed: DbDateTime;
  isactive: Generated<number>;
  isphar: Generated<number>;
  name: string;
  notes: Generated<string | null>;
  version: Generated<string | null>;
}
// ...
export interface SearchTable {
  content: Generated<string | null>;
  object_id: number;
  object_type: string;
  title: Generated<string | null>;
}
// ...
export interface DB {
  _search: SearchTable;
  api_key: ApiKeyTable;
  attachment: AttachmentTable;
  canned_response: CannedResponseTable;
  config: ConfigTable;
  content: ContentTable;
  department: DepartmentTable;
  draft: DraftTable;
  email: EmailTable;
  email_account: EmailAccountTable;
  email_template: EmailTemplateTable;
  email_template_group: EmailTemplateGroupTable;
  event: EventTable;
  faq: FaqTable;
  faq_category: FaqCategoryTable;
  faq_topic: FaqTopicTable;
  file: FileTable;
  file_chunk: FileChunkTable;
  filter: FilterTable;
  filter_action: FilterActionTable;
  filter_rule: FilterRuleTable;
  form: FormTable;
  form_entry: FormEntryTable;
  form_entry_values: FormEntryValuesTable;
  form_field: FormFieldTable;
  group: GroupTable;
  help_topic: HelpTopicTable;
  help_topic_form: HelpTopicFormTable;
  list: ListTable;
  list_items: ListItemsTable;
  lock: LockTable;
  note: NoteTable;
  organization: OrganizationTable;
  organization__cdata: OrganizationCdataTable;
  plugin: PluginTable;
  plugin_instance: PluginInstanceTable;
  queue: QueueTable;
  queue_column: QueueColumnTable;
  queue_columns: QueueColumnsTable;
  queue_config: QueueConfigTable;
  queue_export: QueueExportTable;
  queue_sort: QueueSortTable;
  queue_sorts: QueueSortsTable;
  role: RoleTable;
  schedule: ScheduleTable;
  schedule_entry: ScheduleEntryTable;
  sequence: SequenceTable;
  session: SessionTable;
  sla: SlaTable;
  staff: StaffTable;
  staff_dept_access: StaffDeptAccessTable;
  syslog: SyslogTable;
  task: TaskTable;
  task__cdata: TaskCdataTable;
  team: TeamTable;
  team_member: TeamMemberTable;
  thread: ThreadTable;
  thread_collaborator: ThreadCollaboratorTable;
  thread_entry: ThreadEntryTable;
  thread_entry_email: ThreadEntryEmailTable;
  thread_entry_merge: ThreadEntryMergeTable;
  thread_event: ThreadEventTable;
  thread_referral: ThreadReferralTable;
  ticket: TicketTable;
  ticket__cdata: TicketCdataTable;
  ticket_priority: TicketPriorityTable;
  ticket_status: TicketStatusTable;
  translation: TranslationTable;
  user: UserTable;
  user__cdata: UserCdataTable;
  user_account: UserAccountTable;
  user_email: UserEmailTable;
}